extract steps with conversion metrics to funnel utils
