+++ frontend/src/scenes/funnels/funnelUtils.ts
@@ -9,10 +9,17 @@ import {
     FunnelConversionWindow,
     FunnelsFilterType,
     Breakdown,
+    FunnelStepWithConversionMetrics,
 } from '~/types'
 import { dayjs } from 'lib/dayjs'
 import { combineUrl } from 'kea-router'
 import { FunnelsQuery } from '~/queries/schema'
+
+/* Chosen via heuristics by eyeballing some values
+ * Assuming a normal distribution, then 90% of values are within 1.5 standard deviations of the mean
+ * which gives a ballpark of 1 highlighting every 10 breakdown values
+ */
+const DEVIATION_SIGNIFICANCE_MULTIPLIER = 1.5
 
 const EMPTY_BREAKDOWN_KEY = '__empty_string__'
 const EMPTY_BREAKDOWN_VALUE = '(empty string)'
@@ -314,4 +321,102 @@ export function generateBaselineConversionUrl(url?: string | null): string {
     }
     const parsed = combineUrl(url)
     return combineUrl(parsed.url, { funnel_step_breakdown: undefined }).url
+}
+
+export function stepsWithConversionMetrics(
+    steps: FunnelStepWithNestedBreakdown[],
+    stepReference: FunnelStepReference
+): FunnelStepWithConversionMetrics[] {
+    const stepsWithConversionMetrics = steps.map((step, i) => {
+        const previousCount = i > 0 ? steps[i - 1].count : step.count // previous is faked for the first step
+        const droppedOffFromPrevious = Math.max(previousCount - step.count, 0)
+
+        const nestedBreakdown = step.nested_breakdown?.map((breakdown, breakdownIndex) => {
+            const firstBreakdownCount = steps[0]?.nested_breakdown?.[breakdownIndex].count || 0
+            // firstBreakdownCount serves as previousBreakdownCount for the first step so that
+            // "Relative to previous step" is shown correctly – later series use the actual previous steps
+            const previousBreakdownCount =
+                i === 0 ? firstBreakdownCount : steps[i - 1].nested_breakdown?.[breakdownIndex].count || 0
+            const nestedDroppedOffFromPrevious = Math.max(previousBreakdownCount - breakdown.count, 0)
+            const conversionRates = {
+                fromPrevious: previousBreakdownCount === 0 ? 0 : breakdown.count / previousBreakdownCount,
+                total: breakdown.count / firstBreakdownCount,
+            }
+            return {
+                ...breakdown,
+                droppedOffFromPrevious: nestedDroppedOffFromPrevious,
+                conversionRates: {
+                    ...conversionRates,
+                    fromBasisStep:
+                        stepReference === FunnelStepReference.total
+                            ? conversionRates.total
+                            : conversionRates.fromPrevious,
+                },
+            }
+        })
+
+        const conversionRatesTotal = step.count / steps[0].count
+        const conversionRates = {
+            fromPrevious: previousCount === 0 ? 0 : step.count / previousCount,
+
+            // We get NaN from dividing 0/0 so we just show 0 instead
+            // This is an empty funnel so dropped off percentage will show as 100%
+            // and conversion percentage as 0% but that's better for users than `NaN%`
+            total: Number.isNaN(conversionRatesTotal) ? 0 : conversionRatesTotal,
+        }
+        return {
+            ...step,
+            droppedOffFromPrevious,
+            nested_breakdown: nestedBreakdown,
+            conversionRates: {
+                ...conversionRates,
+                fromBasisStep:
+                    i > 0
+                        ? stepReference === FunnelStepReference.total
+                            ? conversionRates.total
+                            : conversionRates.fromPrevious
+                        : conversionRates.total,
+            },
+        }
+    })
+
+    if (!stepsWithConversionMetrics.length || !stepsWithConversionMetrics[0].nested_breakdown) {
+        return stepsWithConversionMetrics
+    }
+
+    return stepsWithConversionMetrics.map((step) => {
+        // Per step breakdown significance
+        const [meanFromPrevious, stdDevFromPrevious] = getMeanAndStandardDeviation(
+            step.nested_breakdown?.map((item) => item.conversionRates.fromPrevious)
+        )
+        const [meanFromBasis, stdDevFromBasis] = getMeanAndStandardDeviation(
+            step.nested_breakdown?.map((item) => item.conversionRates.fromBasisStep)
+        )
+        const [meanTotal, stdDevTotal] = getMeanAndStandardDeviation(
+            step.nested_breakdown?.map((item) => item.conversionRates.total)
+        )
+
+        const isOutlier = (value: number, mean: number, stdDev: number): boolean => {
+            return (
+                value > mean + stdDev * DEVIATION_SIGNIFICANCE_MULTIPLIER ||
+                value < mean - stdDev * DEVIATION_SIGNIFICANCE_MULTIPLIER
+            )
+        }
+
+        const nestedBreakdown = step.nested_breakdown?.map((item) => {
+            return {
+                ...item,
+                significant: {
+                    fromPrevious: isOutlier(item.conversionRates.fromPrevious, meanFromPrevious, stdDevFromPrevious),
+                    fromBasisStep: isOutlier(item.conversionRates.fromBasisStep, meanFromBasis, stdDevFromBasis),
+                    total: isOutlier(item.conversionRates.total, meanTotal, stdDevTotal),
+                },
+            }
+        })
+
+        return {
+            ...step,
+            nested_breakdown: nestedBreakdown,
+        }
+    })
 }
